pyproject.toml
[project]
name = "vueStockapi"
version = "1.0.0"
description = "Django stock trading backend API with backtesting capabilities"
requires-python = ">=3.10,<3.11"
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14",
    "pandas>=2.0",
    "oracledb>=2.0,<3.0",
    "mysqlclient>=2.2",
    "pymysql>=1.1",
    "celery>=5.3",
    "redis>=5.0",
    "django-celery-beat>=2.5",
    "django-celery-results>=2.5",
    "django-filter>=23.0",
    "django-cors-headers>=4.0",
    "python-decouple>=3.8",
    "setuptools<81",
    "eventlet>=0.33",
    "tushare>=1.2",
    "drf-spectacular>=0.26",
    "backtrader>=1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
    "black>=23.0",
    "flake8>=6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
